% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tccq2_print.R
\name{format.tccq2_type}
\alias{format.tccq2_type}
\title{Format a \code{tccq2} type}
\usage{
\method{format}{tccq2_type}(x, ...)
}
\arguments{
\item{x}{A \code{tccq2_type} object.}

\item{...}{Unused.}
}
\value{
A compact character representation.
}
\description{
Format a \code{tccq2} type
}
