% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tccq_print.R
\name{format.tccq_type}
\alias{format.tccq_type}
\title{Format a \code{tccq} type}
\usage{
\method{format}{tccq_type}(x, ...)
}
\arguments{
\item{x}{A \code{tccq_type} object.}

\item{...}{Unused.}
}
\value{
A compact character representation.
}
\description{
Format a \code{tccq} type
}
