% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tccq2_print.R
\name{print.tccq2_module}
\alias{print.tccq2_module}
\title{Print a \code{tccq2} module summary}
\usage{
\method{print}{tccq2_module}(x, ...)
}
\arguments{
\item{x}{A \code{tccq2_module} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{tccq2} module summary
}
