% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tccq2_print.R
\name{print.tccq2_type}
\alias{print.tccq2_type}
\title{Print a \code{tccq2} type}
\usage{
\method{print}{tccq2_type}(x, ...)
}
\arguments{
\item{x}{A \code{tccq2_type} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{tccq2} type
}
