% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tccq_print.R
\name{print.tccq_module}
\alias{print.tccq_module}
\title{Print a \code{tccq} module summary}
\usage{
\method{print}{tccq_module}(x, ...)
}
\arguments{
\item{x}{A \code{tccq_module} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{tccq} module summary
}
