% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tccq_print.R
\name{print.tccq_type}
\alias{print.tccq_type}
\title{Print a \code{tccq} type}
\usage{
\method{print}{tccq_type}(x, ...)
}
\arguments{
\item{x}{A \code{tccq_type} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{tccq} type
}
