% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_access}
\alias{tccq_access}
\title{Construct a domain access mapping}
\usage{
tccq_access(
  value_id,
  domain,
  kind = "identity",
  index_map = list(),
  attrs = list()
)
}
\arguments{
\item{value_id}{Referenced value id.}

\item{domain}{Access domain.}

\item{kind}{Access kind.}

\item{index_map}{Structured index-map payload.}

\item{attrs}{Structured access attributes.}
}
\description{
Construct a domain access mapping
}
