% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-backend.R
\name{tccq_backend_error_channel}
\alias{tccq_backend_error_channel}
\title{Construct a backend callable error channel}
\usage{
tccq_backend_error_channel(source_name, diagnostics)
}
\arguments{
\item{source_name}{Generated status-output parameter name.}

\item{diagnostics}{Ordered runtime diagnostics addressable by status code.}
}
\description{
Construct a backend callable error channel
}
