% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-backend.R
\name{tccq_backend_extent_binding}
\alias{tccq_backend_extent_binding}
\title{Construct a backend extent binding}
\usage{
tccq_backend_extent_binding(source_name, symbol)
}
\arguments{
\item{source_name}{Generated extent parameter name.}

\item{symbol}{Declared dimension symbol.}
}
\description{
Construct a backend extent binding
}
