% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-backend.R
\name{tccq_backend_plan_set}
\alias{tccq_backend_plan_set}
\title{Construct a backend plan set}
\usage{
tccq_backend_plan_set(
  program_name,
  plans = list(),
  diagnostics = list(),
  attrs = list()
)
}
\arguments{
\item{program_name}{Program name.}

\item{plans}{Backend plans.}

\item{diagnostics}{Diagnostics collected while planning.}

\item{attrs}{Structured plan-set attributes.}
}
\description{
Construct a backend plan set
}
