% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-backend.R
\name{tccq_backend_prepare}
\alias{tccq_backend_prepare}
\title{Prepare a backend plan}
\usage{
tccq_backend_prepare(backend, program, context)
}
\arguments{
\item{backend}{Backend implementation.}

\item{program}{Program to plan.}

\item{context}{Backend planning context.}
}
\description{
Prepare a backend plan
}
