% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_binding}
\alias{tccq_binding}
\title{Construct a program binding}
\usage{
tccq_binding(name, type, mutable = FALSE)
}
\arguments{
\item{name}{Binding name.}

\item{type}{Binding type.}

\item{mutable}{Whether the binding may be mutated.}
}
\description{
Construct a program binding
}
