% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_binding_reference}
\alias{tccq_binding_reference}
\title{Construct a local binding reference}
\usage{
tccq_binding_reference(id, binding)
}
\arguments{
\item{id}{Stable reference value id.}

\item{binding}{Local binding being read.}
}
\description{
Construct a local binding reference
}
