% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_call_index}
\alias{tccq_call_index}
\title{Construct a typed call index}
\usage{
tccq_call_index(calls, semantics = NULL, env = baseenv(), attrs = list())
}
\arguments{
\item{calls}{List of observed calls.}

\item{semantics}{Optional list of call evaluator facts. If omitted, inferred.}

\item{env}{Environment used when inferring missing semantics.}

\item{attrs}{Structured index attributes.}
}
\description{
Construct a typed call index
}
