% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{tccq_cell}
\alias{tccq_cell}
\title{Construct a mutable lexical cell}
\usage{
tccq_cell(name, value_id, type, storage_type = tccq_type(type@base))
}
\arguments{
\item{name}{Binding name.}

\item{value_id}{Stable storage identity.}

\item{type}{Binding type.}

\item{storage_type}{Scalar storage type.}
}
\description{
Construct a mutable lexical cell
}
