% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{tccq_cell_reference}
\alias{tccq_cell_reference}
\title{Construct a lowered mutable-cell read}
\usage{
tccq_cell_reference(id, cell)
}
\arguments{
\item{id}{Stable value id for this read.}

\item{cell}{Mutable cell being read.}
}
\description{
Construct a lowered mutable-cell read
}
