% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_collect_call_semantics}
\alias{tccq_collect_call_semantics}
\title{Collect evaluator facts for calls}
\usage{
tccq_collect_call_semantics(calls, env = baseenv())
}
\arguments{
\item{calls}{List of \code{TccqCall} objects.}

\item{env}{Environment used to resolve ordinary function names.}
}
\description{
Collect evaluator facts for calls
}
