% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_collect_calls}
\alias{tccq_collect_calls}
\title{Collect calls from an R expression}
\usage{
tccq_collect_calls(expr, global_calls = character())
}
\arguments{
\item{expr}{R expression to inspect.}

\item{global_calls}{Additional function names found by \code{codetools}.}
}
\description{
Collect calls from an R expression
}
