% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{tccq_condition_diagnostic}
\alias{tccq_condition_diagnostic}
\title{Extract a diagnostic from a compiler condition}
\usage{
tccq_condition_diagnostic(condition)
}
\arguments{
\item{condition}{A compiler condition.}
}
\description{
Extract a diagnostic from a compiler condition
}
