% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-backend.R
\name{tccq_core_backends}
\alias{tccq_core_backends}
\title{Core backend suite
Core backend descriptors}
\usage{
tccq_core_backends(include_rtinycc = TRUE)
}
\arguments{
\item{include_rtinycc}{Whether to include the \code{Rtinycc} TinyCC descriptor.}
}
\description{
Core backend suite
Core backend descriptors
}
