% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-backend.R
\name{tccq_debug_site}
\alias{tccq_debug_site}
\title{Construct a debug site}
\usage{
tccq_debug_site(id, kind, source = NULL, attrs = list())
}
\arguments{
\item{id}{Stable debug-site id.}

\item{kind}{Debug-site kind.}

\item{source}{Optional source span.}

\item{attrs}{Structured debug-site attributes.}
}
\description{
Construct a debug site
}
