% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_dim_constant}
\alias{tccq_dim_constant}
\title{Construct a constant dimension}
\usage{
tccq_dim_constant(value)
}
\arguments{
\item{value}{Non-negative integer dimension.}
}
\description{
Construct a constant dimension
}
