% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_dim_symbol}
\alias{tccq_dim_symbol}
\title{Construct a symbolic dimension}
\usage{
tccq_dim_symbol(name)
}
\arguments{
\item{name}{Dimension symbol.}
}
\description{
Construct a symbolic dimension
}
