% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_domain}
\alias{tccq_domain}
\title{Construct an iteration domain}
\usage{
tccq_domain(id, shape, axes = NULL, attrs = list())
}
\arguments{
\item{id}{Stable domain id.}

\item{shape}{Domain shape.}

\item{axes}{Optional axis names. Defaults to \code{i1}, \code{i2}, ...}

\item{attrs}{Structured domain attributes.}
}
\description{
Construct an iteration domain
}
