% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_domain_policy}
\alias{tccq_domain_policy}
\title{Construct operation domain policy metadata}
\usage{
tccq_domain_policy(name, result_shape, attrs = list())
}
\arguments{
\item{name}{Human-readable domain policy name.}

\item{result_shape}{Function from input \code{TccqType} list to result
\code{TccqShape}.}

\item{attrs}{Structured domain-policy metadata.}
}
\description{
Construct operation domain policy metadata
}
