% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_domain_policy_result_shape}
\alias{tccq_domain_policy_result_shape}
\title{Return an operation domain-policy result shape}
\usage{
tccq_domain_policy_result_shape(policy, input_types)
}
\arguments{
\item{policy}{Domain policy.}

\item{input_types}{List of input \code{TccqType} values.}
}
\description{
Return an operation domain-policy result shape
}
