% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_elementwise_result_type}
\alias{tccq_elementwise_result_type}
\title{Return an elementwise operation result type}
\usage{
tccq_elementwise_result_type(spec, input_types)
}
\arguments{
\item{spec}{Elementwise metadata.}

\item{input_types}{List of input \code{TccqType} values.}
}
\description{
Return an elementwise operation result type
}
