% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{tccq_expression_tree}
\alias{tccq_expression_tree}
\title{Build a backend-neutral expression tree from a lowered program}
\usage{
tccq_expression_tree(program, value_id = program@result)
}
\arguments{
\item{program}{Lowered program.}

\item{value_id}{Result value id to root the expression tree at.}
}
\description{
Build a backend-neutral expression tree from a lowered program
}
