% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_layout}
\alias{tccq_layout}
\title{Construct a physical array layout}
\usage{
tccq_layout(
  rank,
  order = "unknown",
  strides = NULL,
  offset = tccq_dim_constant(0L),
  contiguous = FALSE
)
}
\arguments{
\item{rank}{Layout rank.}

\item{order}{Layout order.}

\item{strides}{Optional per-axis strides. Defaults to unknown strides.}

\item{offset}{Storage offset.}

\item{contiguous}{Whether the layout is contiguous.}
}
\description{
Construct a physical array layout
}
