% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_literal_finite}
\alias{tccq_literal_finite}
\title{Construct a finite scalar literal}
\usage{
tccq_literal_finite(value, type = NULL)
}
\arguments{
\item{value}{Finite scalar value.}

\item{type}{Optional scalar type. If omitted, it is inferred from \code{value}.}
}
\description{
Construct a finite scalar literal
}
