% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_literal_inf}
\alias{tccq_literal_inf}
\title{Construct an infinite double literal}
\usage{
tccq_literal_inf(sign = 1L)
}
\arguments{
\item{sign}{Positive or negative sign.}
}
\description{
Construct an infinite double literal
}
