% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_literal_na}
\alias{tccq_literal_na}
\title{Construct a typed NA literal}
\usage{
tccq_literal_na(base)
}
\arguments{
\item{base}{Base type for the missing value.}
}
\description{
Construct a typed NA literal
}
