% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_loop_axis}
\alias{tccq_loop_axis}
\title{Construct a loop-nest iteration axis}
\usage{
tccq_loop_axis(name, extent, role = "map")
}
\arguments{
\item{name}{Iteration-axis name.}

\item{extent}{Typed axis extent.}

\item{role}{Axis role, \code{map} or \code{reduce}.}
}
\description{
Construct a loop-nest iteration axis
}
