% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{tccq_loop_transfer}
\alias{tccq_loop_transfer}
\title{Construct a neutral loop transfer}
\usage{
tccq_loop_transfer(id, action, semantics)
}
\arguments{
\item{id}{Stable statement id.}

\item{action}{Loop action, either \code{break} or \code{next}.}

\item{semantics}{Evaluator facts for the originating R control special form.}
}
\description{
Construct a neutral loop transfer
}
