% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_op_registry}
\alias{tccq_op_registry}
\title{Construct an operation registry}
\usage{
tccq_op_registry(implementations = list())
}
\arguments{
\item{implementations}{List of operation implementations.}
}
\description{
Construct an operation registry
}
