% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_op_registry_add}
\alias{tccq_op_registry_add}
\title{Add operation implementations to a registry}
\usage{
tccq_op_registry_add(registry, implementations)
}
\arguments{
\item{registry}{Operation registry.}

\item{implementations}{Operation implementation or list of implementations.}
}
\description{
Add operation implementations to a registry
}
