% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_op_render}
\alias{tccq_op_render}
\title{Render an operation implementation for source output}
\usage{
tccq_op_render(impl, operands, context)
}
\arguments{
\item{impl}{Operation implementation.}

\item{operands}{Rendered operand strings.}

\item{context}{Rendering context.}
}
\description{
Render an operation implementation for source output
}
