% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_op_signature_result_type}
\alias{tccq_op_signature_result_type}
\title{Return an operation signature result type}
\usage{
tccq_op_signature_result_type(signature, input_types)
}
\arguments{
\item{signature}{Operation signature.}

\item{input_types}{List of input \code{TccqType} values.}
}
\description{
Return an operation signature result type
}
