% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_op_supports}
\alias{tccq_op_supports}
\title{Query operation implementation support}
\usage{
tccq_op_supports(impl, call, context)
}
\arguments{
\item{impl}{Operation implementation.}

\item{call}{Observed R call.}

\item{context}{Operation query context.}
}
\description{
Query operation implementation support
}
