% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{tccq_optimize}
\alias{tccq_optimize}
\title{Optimize a typed program}
\usage{
tccq_optimize(optimization, program)
}
\arguments{
\item{optimization}{Optimization implementation.}

\item{program}{Typed program to transform.}
}
\value{
A \link{TccqResult} containing the transformed \link{TccqProgram}.
}
\description{
Optimize a typed program
}
