% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{tccq_pass}
\alias{tccq_pass}
\title{Construct a compiler pass}
\usage{
tccq_pass(name, run)
}
\arguments{
\item{name}{Pass name.}

\item{run}{Pass implementation.}
}
\description{
Construct a compiler pass
}
