% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{tccq_pass_run}
\alias{tccq_pass_run}
\title{Run a compiler pass}
\usage{
tccq_pass_run(pass, program)
}
\arguments{
\item{pass}{Compiler pass.}

\item{program}{Program value.}
}
\description{
Run a compiler pass
}
