% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-backend.R
\name{tccq_plan_backend}
\alias{tccq_plan_backend}
\title{Plan a backend for a typed program}
\usage{
tccq_plan_backend(
  program,
  backend = tccq_c_backend(),
  context = tccq_backend_context()
)
}
\arguments{
\item{program}{Program to plan.}

\item{backend}{Backend implementation.}

\item{context}{Backend planning context.}
}
\description{
Plan a backend for a typed program
}
