% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_reduction_combine}
\alias{tccq_reduction_combine}
\title{Render a reduction combine expression}
\usage{
tccq_reduction_combine(spec, accumulator, value, context)
}
\arguments{
\item{spec}{Reduction metadata.}

\item{accumulator}{Accumulator source expression.}

\item{value}{Current element source expression.}

\item{context}{Operation rendering context.}
}
\description{
Render a reduction combine expression
}
