% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_reduction_identity}
\alias{tccq_reduction_identity}
\title{Return a reducer identity literal}
\usage{
tccq_reduction_identity(spec, type)
}
\arguments{
\item{spec}{Reduction metadata.}

\item{type}{Result type for the reduction.}
}
\description{
Return a reducer identity literal
}
