% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{tccq_reduction_plan}
\alias{tccq_reduction_plan}
\title{Build a neutral reduction plan}
\usage{
tccq_reduction_plan(spec, value, axes, result_type, id, registry)
}
\arguments{
\item{spec}{Reduction metadata.}

\item{value}{Scalar input-element expression.}

\item{axes}{Ordered reduce axes.}

\item{result_type}{Scalar result type.}

\item{id}{Stable loop-nest id.}

\item{registry}{Operation registry used to build transition expressions.}
}
\description{
Build a neutral reduction plan
}
