% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{tccq_reduction_state_component}
\alias{tccq_reduction_state_component}
\title{Construct a reduction-state component}
\usage{
tccq_reduction_state_component(name, target, identity)
}
\arguments{
\item{name}{Semantic component name.}

\item{target}{Neutral local target that stores the component.}

\item{identity}{Initial scalar literal.}
}
\description{
Construct a reduction-state component
}
