% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_region}
\alias{tccq_region}
\title{Construct an executable code region}
\usage{
tccq_region(
  id,
  kind,
  values = list(),
  fusion_groups = list(),
  effect = tccq_effect(),
  memory_space = "host",
  touches_rapi = FALSE,
  attrs = list()
)
}
\arguments{
\item{id}{Stable region id.}

\item{kind}{Region kind: host, kernel, parallel, or device.}

\item{values}{List of IR values in the region.}

\item{fusion_groups}{List of fusion groups in the region.}

\item{effect}{Region effect summary.}

\item{memory_space}{Dominant memory space: r, host, device, or opaque.}

\item{touches_rapi}{Whether the region may call into the R C API.}

\item{attrs}{Structured region attributes.}
}
\description{
Construct an executable code region
}
