% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_registry_supports}
\alias{tccq_registry_supports}
\title{Query whether any registry implementation supports a call}
\usage{
tccq_registry_supports(registry, call, context = tccq_op_context())
}
\arguments{
\item{registry}{Operation registry.}

\item{call}{Observed R call.}

\item{context}{Operation query context.}
}
\description{
Query whether any registry implementation supports a call
}
