% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{tccq_repeat}
\alias{tccq_repeat}
\title{Construct a neutral repeat statement}
\usage{
tccq_repeat(id, body, semantics)
}
\arguments{
\item{id}{Stable statement id.}

\item{body}{Procedural loop body.}

\item{semantics}{Evaluator facts for the originating R loop special form.}
}
\description{
Construct a neutral repeat statement
}
