% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_resolve_call}
\alias{tccq_resolve_call}
\title{Resolve a call to one operation implementation}
\usage{
tccq_resolve_call(registry, call, context = tccq_op_context())
}
\arguments{
\item{registry}{Operation registry.}

\item{call}{Observed R call.}

\item{context}{Operation query context.}
}
\description{
Resolve a call to one operation implementation
}
