% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_resolved_op}
\alias{tccq_resolved_op}
\title{Construct a resolved operation}
\usage{
tccq_resolved_op(call, implementation, attrs = list())
}
\arguments{
\item{call}{Observed R call.}

\item{implementation}{Selected operation implementation.}

\item{attrs}{Structured resolution metadata.}
}
\description{
Construct a resolved operation
}
