% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{tccq_result}
\alias{tccq_result}
\title{Create a compiler result}
\usage{
tccq_result(success, value = NULL, diagnostics = list())
}
\arguments{
\item{success}{Whether the phase completed without diagnostics.}

\item{value}{Result value for the phase.}

\item{diagnostics}{List of \code{TccqDiagnostic} values.}
}
\description{
Create a compiler result
}
