% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{tccq_run_pipeline}
\alias{tccq_run_pipeline}
\title{Run a contract-checked pass pipeline}
\usage{
tccq_run_pipeline(program, passes)
}
\arguments{
\item{program}{Program value.}

\item{passes}{Compiler pass or list of passes.}
}
\description{
Run a contract-checked pass pipeline
}
