% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_shape}
\alias{tccq_shape}
\title{Construct a shape}
\usage{
tccq_shape(dims = list())
}
\arguments{
\item{dims}{A list, character vector, integer vector, or single dimension.}
}
\description{
Construct a shape
}
