% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-backend.R
\name{tccq_source_span}
\alias{tccq_source_span}
\title{Construct a source span}
\usage{
tccq_source_span(
  file = "",
  line = NA_integer_,
  column = NA_integer_,
  end_line = NA_integer_,
  end_column = NA_integer_,
  label = ""
)
}
\arguments{
\item{file}{Source file, or empty string if unknown.}

\item{line}{Starting line, or \code{NA_integer_} if unknown.}

\item{column}{Starting column, or \code{NA_integer_} if unknown.}

\item{end_line}{Ending line, or \code{NA_integer_} if unknown.}

\item{end_column}{Ending column, or \code{NA_integer_} if unknown.}

\item{label}{Human-readable source label.}
}
\description{
Construct a source span
}
