% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_subscript_spec}
\alias{tccq_subscript_spec}
\title{Construct subscript implementation metadata}
\usage{
tccq_subscript_spec(name, signature)
}
\arguments{
\item{name}{Human-readable subscript operation name.}

\item{signature}{Shared operation signature.}
}
\description{
Construct subscript implementation metadata
}
