% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_tile}
\alias{tccq_tile}
\title{Construct tile metadata}
\usage{
tccq_tile(shape, origin = NULL)
}
\arguments{
\item{shape}{Tile shape.}

\item{origin}{Optional per-axis origin. Defaults to zero origin.}
}
\description{
Construct tile metadata
}
