% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_unsupported_calls}
\alias{tccq_unsupported_calls}
\title{Find unsupported calls for an operation registry and context}
\usage{
tccq_unsupported_calls(
  calls,
  registry = tccq_default_op_registry(),
  context = tccq_op_context()
)
}
\arguments{
\item{calls}{List of \code{TccqCall} objects.}

\item{registry}{Operation registry.}

\item{context}{Operation query context.}
}
\description{
Find unsupported calls for an operation registry and context
}
