% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_value}
\alias{tccq_value}
\title{Construct an IR value}
\usage{
tccq_value(
  id,
  op,
  inputs = list(),
  type,
  effect = tccq_effect(),
  layout = NULL,
  tile = NULL,
  attrs = list()
)
}
\arguments{
\item{id}{Stable value id.}

\item{op}{Operation name.}

\item{inputs}{Input value ids or references.}

\item{type}{Result type.}

\item{effect}{Effect summary.}

\item{layout}{Optional physical layout.}

\item{tile}{Optional tile metadata.}

\item{attrs}{Structured operation attributes.}
}
\description{
Construct an IR value
}
