% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{TccqAssignment}
\alias{TccqAssignment}
\title{Neutral assignment statement}
\usage{
TccqAssignment(
  id = character(0),
  effect = TccqEffect(),
  target = TccqWriteTarget(),
  value = TccqExpression()
)
}
\arguments{
\item{id}{Stable statement id.}

\item{effect}{Effect summary of evaluating the statement.}

\item{target}{Typed destination.}

\item{value}{Backend-neutral value expression.}
}
\description{
Neutral assignment statement
}
