% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-backend.R
\docType{data}
\name{TccqBackend}
\alias{TccqBackend}
\title{Backend implementation trait}
\format{
An object of class \code{s7contract::s7_trait} (inherits from \code{S7_object}) of length 1.
}
\usage{
TccqBackend
}
\description{
Backend implementations must opt into this trait before compiler code uses
them for planning.
}
\keyword{datasets}
