% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{TccqBinding}
\alias{TccqBinding}
\title{Program binding}
\usage{
TccqBinding(name = character(0), type = TccqType(), mutable = logical(0))
}
\arguments{
\item{name}{Binding name.}

\item{type}{Binding type.}

\item{mutable}{Whether the binding may be mutated.}
}
\description{
Program binding
}
