% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{TccqCellReference}
\alias{TccqCellReference}
\title{Lowered mutable-cell read}
\usage{
TccqCellReference(
  id = character(0),
  op = character(0),
  inputs = list(),
  type = TccqType(),
  effect = TccqEffect(),
  attrs = list(),
  cell = TccqCell()
)
}
\arguments{
\item{id}{Stable value id.}

\item{op}{Operation name.}

\item{inputs}{Input value ids or references.}

\item{type}{Result type.}

\item{effect}{Effect summary.}

\item{attrs}{Structured operation attributes.}

\item{cell}{Mutable cell being read.}
}
\description{
Lowered mutable-cell read
}
