% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{TccqConditional}
\alias{TccqConditional}
\title{Neutral conditional statement}
\usage{
TccqConditional(
  id = character(0),
  effect = TccqEffect(),
  condition = TccqExpression(),
  consequent = TccqValueBlock(),
  alternative = TccqValueBlock(),
  branch = TccqBranch()
)
}
\arguments{
\item{id}{Stable statement id.}

\item{effect}{Effect summary of evaluating the statement.}

\item{condition}{Scalar logical condition expression.}

\item{consequent}{Block evaluated when the condition is true.}

\item{alternative}{Block evaluated when the condition is false.}

\item{branch}{Source branch payload retaining R special-form semantics.}
}
\description{
Neutral conditional statement
}
