% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-backend.R
\name{TccqDebugSite}
\alias{TccqDebugSite}
\title{Debug site in an IR or backend plan}
\usage{
TccqDebugSite(
  id = character(0),
  kind = character(0),
  source = NULL,
  attrs = list()
)
}
\arguments{
\item{id}{Stable debug-site id.}

\item{kind}{Debug-site kind.}

\item{source}{Optional source span.}

\item{attrs}{Structured debug-site attributes.}
}
\description{
Debug site in an IR or backend plan
}
