% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{TccqDim}
\alias{TccqDim}
\title{Shape dimension}
\usage{
TccqDim(kind = character(0), label = character(0), value = integer(0))
}
\arguments{
\item{kind}{Dimension kind.}

\item{label}{Symbolic dimension label.}

\item{value}{Constant dimension value.}
}
\description{
Shape dimension
}
