% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{TccqIf}
\alias{TccqIf}
\title{Neutral procedural if statement}
\usage{
TccqIf(
  id = character(0),
  effect = TccqEffect(),
  condition = TccqExpression(),
  consequent = TccqBlock(),
  alternative = TccqBlock(),
  semantics = TccqCallSemantics()
)
}
\arguments{
\item{id}{Stable statement id.}

\item{effect}{Effect summary of evaluating the statement.}

\item{condition}{Scalar logical condition expression.}

\item{consequent}{Block evaluated when the condition is true.}

\item{alternative}{Block evaluated when the condition is false.}

\item{semantics}{Evaluator facts for the originating \code{if} special form.}
}
\description{
Neutral procedural if statement
}
