% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{TccqLiteral}
\alias{TccqLiteral}
\title{Scalar literal value}
\usage{
TccqLiteral(kind = character(0), type = TccqType(), value = NULL)
}
\arguments{
\item{kind}{Literal kind.}

\item{type}{Scalar type.}

\item{value}{Literal payload.}
}
\description{
Scalar literal value
}
