% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{TccqOpImplementation}
\alias{TccqOpImplementation}
\title{Operation implementation trait}
\usage{
TccqOpImplementation
}
\description{
Implementations must explicitly opt into this trait before they can be used
by an operation registry.
}
