% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{TccqOpRenderContext}
\alias{TccqOpRenderContext}
\title{Operation source-rendering context}
\usage{
TccqOpRenderContext(
  language = character(0),
  backend_id = character(0),
  attrs = list()
)
}
\arguments{
\item{language}{Target source language.}

\item{backend_id}{Backend requesting rendering.}

\item{attrs}{Structured rendering metadata.}
}
\description{
Operation source-rendering context
}
