% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\docType{data}
\name{TccqPass}
\alias{TccqPass}
\title{Compiler pass interface}
\format{
An object of class \code{s7contract::s7_interface} (inherits from \code{S7_object}) of length 1.
}
\usage{
TccqPass
}
\description{
Compiler pass interface
}
\keyword{datasets}
