% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{TccqPassSpec}
\alias{TccqPassSpec}
\title{Compiler pass implementation}
\usage{
TccqPassSpec(name = character(0), run = function() NULL)
}
\arguments{
\item{name}{Pass name.}

\item{run}{Pass implementation.}
}
\description{
A pass is an S7 object whose \code{run} function must accept a \code{TccqProgram} and
return a \code{TccqProgram}. The protocol is also represented as a \code{s7contract}
interface so passes can be checked progressively at runtime.
}
