% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{TccqReductionState}
\alias{TccqReductionState}
\title{Typed neutral reduction state}
\usage{
TccqReductionState(components = list())
}
\arguments{
\item{components}{Named reduction-state components.}
}
\description{
Typed neutral reduction state
}
