% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{TccqRegion}
\alias{TccqRegion}
\title{Executable code region}
\usage{
TccqRegion(
  id = character(0),
  kind = character(0),
  values = list(),
  fusion_groups = list(),
  effect = TccqEffect(),
  memory_space = character(0),
  touches_rapi = logical(0),
  attrs = list()
)
}
\arguments{
\item{id}{Stable region id.}

\item{kind}{Region kind.}

\item{values}{List of IR values in the region.}

\item{fusion_groups}{List of fusion groups in the region.}

\item{effect}{Region effect summary.}

\item{memory_space}{Dominant memory space for the region.}

\item{touches_rapi}{Whether the region may call into the R C API.}

\item{attrs}{Structured region attributes.}
}
\description{
Regions classify a group of IR values by where they may run and whether they
may touch the R C API. \code{kernel}, \code{parallel}, and \code{device} regions are
intended to be R-API-free.
}
