% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{TccqShape}
\alias{TccqShape}
\title{Shape value}
\usage{
TccqShape(rank = integer(0), dims = list())
}
\arguments{
\item{rank}{Number of dimensions.}

\item{dims}{List of \code{TccqDim} values.}
}
\description{
Shape value
}
