% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-backend.R
\name{TccqSourceSpan}
\alias{TccqSourceSpan}
\title{Source span for IR/debug metadata}
\usage{
TccqSourceSpan(
  file = character(0),
  line = integer(0),
  column = integer(0),
  end_line = integer(0),
  end_column = integer(0),
  label = character(0)
)
}
\arguments{
\item{file}{Source file, or empty string if unknown.}

\item{line}{Starting line, or \code{NA_integer_} if unknown.}

\item{column}{Starting column, or \code{NA_integer_} if unknown.}

\item{end_line}{Ending line, or \code{NA_integer_} if unknown.}

\item{end_column}{Ending column, or \code{NA_integer_} if unknown.}

\item{label}{Human-readable source label.}
}
\description{
Source span for IR/debug metadata
}
