% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{TccqStoragePlan}
\alias{TccqStoragePlan}
\title{Storage plan}
\usage{
TccqStoragePlan(slots = list())
}
\arguments{
\item{slots}{List of storage slots.}
}
\description{
A storage plan belongs to the middle end. It is consumed by backends but does
not depend on C, Fortran, TinyCC, or any concrete emitter.
}
