% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{TccqTile}
\alias{TccqTile}
\title{Array tile description}
\usage{
TccqTile(shape = TccqShape(), origin = list())
}
\arguments{
\item{shape}{Tile shape.}

\item{origin}{Per-axis tile origin.}
}
\description{
Tile metadata describes a rectangular partition of an array domain. It does
not change the value type.
}
