% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{TccqValue}
\alias{TccqValue}
\title{IR value}
\usage{
TccqValue(
  id = character(0),
  op = character(0),
  inputs = list(),
  type = TccqType(),
  effect = TccqEffect(),
  attrs = list()
)
}
\arguments{
\item{id}{Stable value id.}

\item{op}{Operation name.}

\item{inputs}{Input value ids or references.}

\item{type}{Result type.}

\item{effect}{Effect summary.}

\item{attrs}{Structured operation attributes.}
}
\description{
IR value
}
