% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{tccq_assignment}
\alias{tccq_assignment}
\title{Construct a neutral assignment}
\usage{
tccq_assignment(id, target, value)
}
\arguments{
\item{id}{Stable statement id.}

\item{target}{Typed destination.}

\item{value}{Backend-neutral value expression.}
}
\description{
Construct a neutral assignment
}
