% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-expression.R
\name{tccq_block}
\alias{tccq_block}
\title{Construct a neutral statement block}
\usage{
tccq_block(id, locals = list(), statements = list(), effect = tccq_effect())
}
\arguments{
\item{id}{Stable block id.}

\item{locals}{Ordered scalar local write targets declared by the block.}

\item{statements}{Ordered \code{TccqStatement} values.}

\item{effect}{Effect summary of evaluating the block.}
}
\description{
Construct a neutral statement block
}
