% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_call}
\alias{tccq_call}
\title{Construct an observed call}
\usage{
tccq_call(name, expr = NULL, origin = "ast")
}
\arguments{
\item{name}{Call name.}

\item{expr}{Original R call expression.}

\item{origin}{Source of the call observation.}
}
\description{
Construct an observed call
}
