% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_call}
\alias{tccq_call}
\title{Construct an observed call}
\usage{
tccq_call(
  name,
  expr = NULL,
  origin = "ast",
  kind = NULL,
  arity = NULL,
  argument_names = NULL,
  attrs = list()
)
}
\arguments{
\item{name}{Call name.}

\item{expr}{Original R call expression.}

\item{origin}{Source of the call observation.}

\item{kind}{Structural call kind.}

\item{arity}{Number of supplied arguments, or \code{NA_integer_}.}

\item{argument_names}{Supplied argument tags.}

\item{attrs}{Structured call attributes.}
}
\description{
Construct an observed call
}
