% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_default_op_registry}
\alias{tccq_default_op_registry}
\title{Default operation registry for the reset frontend}
\usage{
tccq_default_op_registry()
}
\description{
Default operation registry for the reset frontend
}
