% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tccq_op_signature}
\alias{tccq_op_signature}
\title{Construct operation signature metadata}
\usage{
tccq_op_signature(name, arity, result_type, attrs = list())
}
\arguments{
\item{name}{Human-readable operation signature name.}

\item{arity}{Accepted argument counts.}

\item{result_type}{Function from input \code{TccqType} list to result \code{TccqType}.}

\item{attrs}{Structured signature metadata.}
}
\description{
Construct operation signature metadata
}
