% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-backend.R
\name{tccq_plan_backends}
\alias{tccq_plan_backends}
\title{Plan several backends for a typed program}
\usage{
tccq_plan_backends(
  program,
  backends = tccq_core_backends(),
  context = tccq_backend_context()
)
}
\arguments{
\item{program}{Program to plan.}

\item{backends}{Backend implementation descriptors.}

\item{context}{Backend planning context.}
}
\description{
Plan several backends for a typed program
}
