% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_program}
\alias{tccq_program}
\title{Construct a program schema}
\usage{
tccq_program(
  name,
  formals,
  values = list(),
  result = NULL,
  diagnostics = list()
)
}
\arguments{
\item{name}{Program name.}

\item{formals}{Named list of formal bindings.}

\item{values}{List of IR values.}

\item{result}{Result value id or object.}

\item{diagnostics}{List of diagnostics attached to the program.}
}
\description{
Construct a program schema
}
