% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_storage_plan}
\alias{tccq_storage_plan}
\title{Construct a storage plan}
\usage{
tccq_storage_plan(slots = list())
}
\arguments{
\item{slots}{List of storage slots.}
}
\description{
Construct a storage plan
}
