% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{tccq_type}
\alias{tccq_type}
\title{Construct a compiler type}
\usage{
tccq_type(base, shape = tccq_shape())
}
\arguments{
\item{base}{Scalar or storage base type.}

\item{shape}{Shape value.}
}
\description{
Construct a compiler type
}
