% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{TccqCall}
\alias{TccqCall}
\title{R call observed by the frontend}
\usage{
TccqCall(name = character(0), expr = NULL, origin = character(0))
}
\arguments{
\item{name}{Call name.}

\item{expr}{Original R call expression.}

\item{origin}{Source of the call observation.}
}
\description{
R call observed by the frontend
}
