% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{data}
\name{TccqOpImplementation}
\alias{TccqOpImplementation}
\title{Operation implementation trait}
\format{
An object of class \code{s7contract::s7_trait} (inherits from \code{S7_object}) of length 1.
}
\usage{
TccqOpImplementation
}
\description{
Implementations must explicitly opt into this trait before they can be used
by an operation registry.
}
\keyword{datasets}
