% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{TccqOpRegistry}
\alias{TccqOpRegistry}
\title{Operation registry}
\usage{
TccqOpRegistry(implementations = list())
}
\arguments{
\item{implementations}{List of operation implementations.}
}
\description{
Operation registry
}
