% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{TccqProgram}
\alias{TccqProgram}
\title{Program schema}
\usage{
TccqProgram(
  values = list(),
  result = NULL,
  name = character(0),
  formals = list(),
  schedule = NULL,
  regions = list(),
  diagnostics = list(),
  call_index = NULL,
  storage_plan = NULL,
  attrs = list()
)
}
\arguments{
\item{values}{List of IR values.}

\item{result}{Result value id or object.}

\item{name}{Program name.}

\item{formals}{Named list of formal bindings.}

\item{schedule}{Optional ordered program evaluation schedule.}

\item{regions}{List of executable code regions.}

\item{diagnostics}{List of diagnostics attached to the program.}

\item{call_index}{Typed frontend call handoff.}

\item{storage_plan}{Optional middle-end storage plan.}

\item{attrs}{Structured program attributes.}
}
\description{
Program schema
}
