% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{TccqProgram}
\alias{TccqProgram}
\title{Program schema}
\usage{
TccqProgram(
  name = character(0),
  formals = list(),
  values = list(),
  regions = list(),
  result = NULL,
  diagnostics = list()
)
}
\arguments{
\item{name}{Program name.}

\item{formals}{Named list of formal bindings.}

\item{values}{List of IR values.}

\item{regions}{List of executable code regions.}

\item{result}{Result value id or object.}

\item{diagnostics}{List of diagnostics attached to the program.}
}
\description{
Program schema
}
