% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{TccqResult}
\alias{TccqResult}
\title{Compiler result value}
\usage{
TccqResult(ok = logical(0), value = NULL, diagnostics = list())
}
\arguments{
\item{ok}{Whether compilation can continue.}

\item{value}{Result value for the phase.}

\item{diagnostics}{List of \code{TccqDiagnostic} values.}
}
\description{
Compiler result value
}
