% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-schema.R
\name{TccqType}
\alias{TccqType}
\title{Declared compiler type}
\usage{
TccqType(base = character(0), shape = TccqShape())
}
\arguments{
\item{base}{Scalar or storage base type.}

\item{shape}{Shape value.}
}
\description{
Declared compiler type
}
