% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tccquickr-package}
\alias{tccquickr}
\alias{tccquickr-package}
\title{tccquickr: typed experimental R transformation core}
\description{
\code{tccquickr} is being rebuilt around declared R programs, shape-aware types,
classed diagnostics, contract-checked compiler passes, and generic backend
planning. Rtinycc is modeled as one backend descriptor, not as the compiler
architecture.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sounkou-bioinfo/tccquickr}
  \item Report bugs at \url{https://github.com/sounkou-bioinfo/tccquickr/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

}
\keyword{internal}
